Start a new file for future preprint.
I think it is best to proceed like that instead of trying to modify an
merge directly both WIP-files that we have. I take the opportunity to
rewrite completely the section about the circle in that new file, in
the spirit of what I would have presented at TYPES. On one hand, it is
probably more ppedagogical and conceptually clear in this newer
version than in the older. On the other hand, the actual formula for
the equivalence (S1+S1) = (S1=S1) is never fully displayed. I am not
sure which is the best of the two.